Word Count - VBA Macro
Many times we have to deal with “keywords” while working on texts, strings.; and we need to know which words occur often and which do not.  So anything handy to give us this information is always a “good to have” thingy and it is always easy to do this using some programming languages (R, Python, etc.).  However I was looking for something to be quicker and handier especially when working on Excel files – and nothing comes as handy as a VBA macro
The macro is just a basic extraction of words and its frequency of occurrence